Create hBot
This is a simple bot that uses heuristics and hard coding, but no search. It is not sufficient for what we need to do. However, it outperforms RaisedPlayer and RandomPlayer. Here are the results of running testperf with 500 gmaes of 1000 rounds:
hBot's final pot:  9,347,420
Raised's final pot:  650,240
hBot's final pot:  9,997,380
Random's final pot:  0
Therefore, this bot will be a nice benchmark for us to test our MCTSBot against.